sudden_infant_death/requirements.txt
inference-sdk==0.2.8
python-dotenv==1.0.1
opencv-python==4.8.1.78